+++ package.json
@@ -9,7 +9,6 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@unoCSS/transformer-directives": "npm:@unocss/transformer-directives@^0.49.4",
     "@vitejs/plugin-vue-jsx": "^3.0.0",
     "@vueuse/core": "^9.12.0",
     "@vueuse/integrations": "^9.12.0",
@@ -28,10 +27,10 @@
   },
   "devDependencies": {
     "@iconify/json": "^2.2.18",
+    "@iconify/vue": "^4.1.0",
     "@vicons/ionicons5": "^0.12.0",
     "@vitejs/plugin-vue": "^4.0.0",
     "naive-ui": "^2.34.3",
-    "unocss": "^0.49.4",
     "unplugin-auto-import": "^0.12.1",
     "unplugin-icons": "^0.15.2",
     "unplugin-vue-components": "^0.22.12",
