+++ src/assets/styles/index.scss
@@ -5,14 +5,12 @@
 }
 html,
 body {
+  font-size: 16px;
   width: 100%;
   height: 100%;
   overflow: hidden;
 }
 
-html {
-  font-size: 4px; // * 1rem = 4px  方便unocss计算：在unocss中 1字体单位 = 0.25rem，相当于 1等份 = 1px
-}
 
 body {
   font-size: 16px;
