+++ src/layout/components/sidebar/index.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    sidebar
+  </div>
+</template>
+
+<script setup>
+
+</script>
+
+<style lang="scss" scoped>
+
+</style>
