+++ src/layout/components/tags/index.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    tags
+  </div>
+</template>
+
+<script setup>
+
+</script>
+
+<style lang="scss" scoped>
+
+</style>
