--- src/layouts/components/sidebar/index.vue
@@ -1,9 +1,0 @@
-<script setup>
-import SideLogo from './components/SideLogo.vue'
-import SideMenu from './components/SideMenu.vue'
-</script>
-
-<template>
-  <SideLogo />
-  <SideMenu />
-</template>
