+++ src/utils/icon.js
@@ -1,12 +1,8 @@
 import { h } from 'vue';
 import { Icon } from '@iconify/vue';
 import { NIcon } from 'naive-ui';
-// import SvgIcon from '@/components/icon/SvgIcon.vue'
 
 export function renderIcon(icon, props = { size: 12 }) {
   return () => h(NIcon, props, { default: () => h(Icon, { icon }) })
 }
 
-export function renderCustomIcon(icon, props = { size: 12 }) {
-  return () => h(NIcon, props, { default: () => h(Icon, { icon }) })
-}
